+++ ImageAnalysis/image_analysis/analyzers/calibration_scripts/modules_image_processing/shot_charge_reader.py
